reducer테스트를 RestaurantSlice테스트로 옮겨라
